+++ NEWS.md
@@ -6,11 +6,11 @@ output: html_document
 ---
 
 ## DTedit 2.2.4
-21st October 2020
+23rd October 2020
 
 ### New
 
-* `shinyFeedback` parameter to use https://github.com/merlinoa/shinyFeedback
+* `inputEvent` parameter to use https://github.com/merlinoa/shinyFeedback
    (issue #17 and issue #9)
 
 ### Bug fix